requirements.txt
asn1crypto==0.24.0
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
click==6.7
cryptography==2.1.4
ecdsa==0.13
Flask==1.0.2
Flask-Assets==0.12
Flask-Caching==1.4.0
Flask-Login==0.4.1
Flask-OAuthlib==0.9.4
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
future==0.16.0
gunicorn==19.7.1
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
marshmallow==2.15.0
marshmallow-jsonapi==0.16.0
oauthlib==2.0.6
packaging==16.8
pip-check-reqs==2.0.1
pycparser==2.18
pycryptodome==3.5.1
pyparsing==2.2.0
python-dateutil==2.6.1
python-dotenv==0.8.2
python-jose==2.0.2
redis==2.10.6
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
SQLAlchemy==1.2.0
urllib3==1.22
webassets==0.12.1
Werkzeug==0.14.1
WTForms==2.1

